# docs/swagger.yaml
basePath: /
definitions:
  main.Activity:
    description: 配對活動資訊
    properties:
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      location:
        $ref: '#/definitions/main.Location'
      location_id:
        type: integer
      target_count:
        type: integer
      title:
        type: string
    type: object
...
  main.Location:
    description: 地點資訊
    properties:
      address:
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
  main.Match:
    description: 配對局資訊
    properties:
      activity:
        $ref: '#/definitions/main.Activity'
      activity_id:
        type: integer
      id:
        type: integer
      match_time:
        type: string
      organizer:
        $ref: '#/definitions/main.User'
      organizer_id:
        type: integer
      status:
        description: open, closed, completed
        type: string
    type: object
  main.MatchParticipant:
    description: 配對參與者資訊
    properties:
      id:
        type: integer
      joined_at:
        type: string
      match:
        $ref: '#/definitions/main.Match'
      match_id:
        type: integer
      status:
        description: pending, approved, rejected
        type: string
      user:
        $ref: '#/definitions/main.User'
      user_id:
        type: integer
    type: object
...
  main.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: integer
      email:
        type: string
      id:
        type: integer
      is_admin:
        description: 添加管理員標記
        type: boolean
      name:
        type: string
      social_id:
        type: string
      social_provider:
        type: string
      updated_at:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 這是一個買一送一配對網站的API文檔
  title: 買一送一配對網站 API
  version: "1.0"
paths:
...
  /user/matches:
    get:
      consumes:
      - application/json
      description: 取得所有時間未到且狀態為open的配對列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Match'
            type: array
        "500":
          description: 無法取得配對列表
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 取得時間未到的配對列表
      tags:
      - 使用者
    post:
      consumes:
      - application/json
      description: 建立新的配對局 (開局)
      parameters:
      - description: 配對局資訊
        in: body
        name: match
        required: true
        schema:
          $ref: '#/definitions/main.Match'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Match'
        "400":
          description: 無效的請求資料
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法建立配對局
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 建立新的配對局
      tags:
      - 使用者
  /user/matches/{id}/join:
    post:
      consumes:
      - application/json
      description: 參與指定ID的配對局
      parameters:
      - description: 配對局ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.MatchParticipant'
        "400":
          description: 無效的配對局 ID 或已參與
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法參與配對局
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 參與配對
      tags:
      - 使用者
  /user/past-matches:
    get:
      consumes:
      - application/json
      description: 取得該使用者參與過的已完成的配對局列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Match'
            type: array
        "500":
          description: 無法取得過去參與的配對列表
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 取得過去參與的配對列表
      tags:
      - 使用者
securityDefinitions:
  ApiKeyAuth:
    description: 輸入 'Bearer &lt;JWT token&gt;' 進行認證。可以先透過 Facebook 登入取得 token。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"